[v0x04] Moving instructions from flow_match to flow_instructions
