+++ pyof/v0x04/common/flow_match.py
@@ -1,39 +1,15 @@
-"""Defines flow statistics structures and related items.
-
-An OpenFlow match is composed of a flow match header and a sequence of zero or
-more flow match fields
-"""
+"""An OpenFlow match is composed of a flow match header and a sequence of zero
+or more flow match fields"""
 # System imports
 from enum import Enum
 # Third-party imports
 
 # Local source tree imports
 from pyof.foundation.base import GenericStruct
-from pyof.foundation.basic_types import Pad, UBInt8, UBInt16, UBInt32, UBInt64
-from pyof.v0x04.common.action import ListOfActions
+from pyof.foundation.basic_types import UBInt8, UBInt16, UBInt32
 
-__all__ = ('InstructionApplyAction', 'InstructionClearAction',
-           'InstructionGotoTable', 'InstructionMeter', 'InstructionType',
-           'InstructionWriteAction', 'InstructionWriteMetadata',
-           'Ipv6ExtHdrFlags', 'Match', 'MatchField', 'MatchType',
+__all__ = ('Ipv6ExtHdrFlags', 'Match', 'MatchField', 'MatchType',
            'OxmExperimenterHeader', 'OxmOfbMatchField', 'VlanId')
-
-class InstructionType(Enum):
-    """List of instructions that are currently defined."""
-    #: Setup the next table in the lookup pipeline
-    OFPIT_GOTO_TABLE = 1
-    #: Setup the metadata field for use later in pipeline
-    OFPIT_WRITE_METADATA = 2
-    #: Write the action(s) onto the datapath action set
-    OFPIT_WRITE_ACTIONS = 3
-    #: Applies the action(s) immediately
-    OFPIT_APPLY_ACTIONS = 4
-    #: Clears all actions from the datapath action set
-    OFPIT_CLEAR_ACTIONS = 5
-    #: Apply meter (rate limiter)
-    OFPIT_METER = 6
-    #: Experimenter instruction
-    OFPIT_EXPERIMENTER = 0xFFFF
 
 
 class Ipv6ExtHdrFlags(Enum):
@@ -193,136 +169,6 @@ class VlanId(Enum):
 # Classes
 
 
-class InstructionApplyAction(GenericStruct):
-    """Instruction structure for OFPIT_APPLY_ACTIONS.
-
-    The :attr:`~actions` field is treated as a list, and the actions are
-    applied to the packet in-order
-    """
-    #: OFPIT_APPLY_ACTIONS
-    instruction_type = UBInt16(InstructionType.OFPIT_APPLY_ACTIONS)
-    #: Length of this struct in bytes.
-    length = UBInt16()
-    #: Align to 64-bits
-    pad = Pad(4)
-    #: Actions associated with OFPIT_APPLY_ACTIONS
-    actions = ListOfActions()
-
-    def __init__(self, length=None, actions=None):
-        """
-        Args:
-            - length (int): Length of this struct in bytes
-            - actions (:class:`~.actions.ListOfActions`): Actions associated
-                with OFPIT_APPLY_ACTIONS
-        """
-        self.length = length
-        self.actions = actions if actions is not None else []
-
-
-class InstructionClearAction(GenericStruct):
-    """Instruction structure for OFPIT_CLEAR_ACTIONS.
-
-    This structure does not contain any actions"""
-    #: OFPIT_CLEAR_ACTIONS
-    instruction_type = UBInt16(InstructionType.OFPIT_CLEAR_ACTIONS)
-    #: Length of this struct in bytes.
-    length = UBInt16(8)
-    #: Align to 64-bits
-    pad = Pad(4)
-    #: OFPIT_CLEAR_ACTIONS does not have any action on the list of actions.
-    actions = ListOfActions()
-
-
-class InstructionGotoTable(GenericStruct):
-    """Instruction structure for OFPIT_GOTO_TABLE."""
-    #: OFPIT_GOTO_TABLE.
-    instruction_type = UBInt16(InstructionType.OFPIT_GOTO_TABLE)
-    #: Length of this struct in bytes.
-    length = UBInt16()
-    #: Set next table in the lookup pipeline.
-    table_id = UBInt8()
-    #: Pad to 64 bits
-    pad = Pad(3)
-
-    def __init__(self, length=None, table_id=None):
-        """
-        Args:
-            - length (int): Length of this struct in bytes
-            - table_id (int): set next table in the lookup pipeline
-        """
-        self.length = length
-        self.table_id = table_id
-
-
-class InstructionMeter(GenericStruct):
-    """Instruction structure for OFPIT_METER.
-
-    meter_id indicates which meter to apply on the packet."""
-    #: OFPIT_METER.
-    instruction_type = UBInt16(InstructionType.OFPIT_METER)
-    #: Length is 8.
-    length = UBInt16(8)
-    #: Meter instance.
-    meter_id = UBInt32()
-
-    def __init__(self, meter_id=None):
-        """
-        Args:
-            - meter_id (int): Meter instance
-        """
-        self.meter_id = meter_id
-
-
-class InstructionWriteAction(GenericStruct):
-    """Instruction structure for OFPIT_WRITE_ACTIONS.
-
-    The actions field must be treated as a SET, so the actions are not repeated
-    """
-    #: OFPIT_WRITE_ACTIONS
-    instruction_type = UBInt16(InstructionType.OFPIT_WRITE_ACTIONS)
-    #: Length of this struct in bytes.
-    length = UBInt16()
-    #: Align to 64-bits
-    pad = Pad(4)
-    #: Actions associated with OFPIT_WRITE_ACTIONS
-    actions = ListOfActions()
-
-    def __init__(self, length=None, actions=None):
-        """
-        Args:
-            - length (int): Length of this struct in bytes
-            - actions (:class:`~.actions.ListOfActions`): Actions associated
-                with OFPIT_WRITE_ACTIONS
-        """
-        self.length = length
-        self.actions = actions if actions is not None else []
-
-
-class InstructionWriteMetadata(GenericStruct):
-    """Instruction structure for OFPIT_WRITE_METADATA."""
-    #: OFPIT_WRITE_METADATA
-    instruction_type = UBInt16(InstructionType.OFPIT_WRITE_METADATA)
-    #: Length of this struct in bytes
-    length = UBInt16()
-    #: Align to 64-bits
-    pad = Pad(4)
-    #: Metadata value to write
-    metadata = UBInt64()
-    #: Metadata write bitmask
-    metadata_mask = UBInt64()
-
-    def __init__(self, length=None, metadata=None, metadata_mask=None):
-        """
-        Args:
-            - length (int): Length of this struct in bytes
-            - metadata (int): Metadata value to write
-            - metadata_mask (int): Metadata write bitmask
-        """
-        self.length = length
-        self.metadata = metadata
-        self.metadata_mask = metadata_mask
-
-
 class Match(GenericStruct):
     """Describes the flow match header structure.
 
@@ -383,4 +229,5 @@ class OxmExperimenterHeader(GenericStruct):
             - experimenter (int): Experimenter ID which takes the same form as
                 in struct ofp_experimenter_header
         """
-        self.Experimenter = experimenter
+        super().__init__()
+        self.experimenter = experimenter
